ARM: mvebu: Warn about the wake-ups sources not taken into account in suspend
On the Armada 370/XP/38x/39x SoCs when the suspend to ram feature is
supported, the SoC is shutdown and will be wake up by an external
micro-controller, so there is no possibility to setup wake-up source
from Linux. However, in standby mode, as the SoC stay powered and it is
possible to wake-up from any interrupt sources. As when user decide to
setup a wake-up source, there is no way to know if they are wake-up
source from suspend or from standby, then we chose allowing to setup
all the interrupt as wake-up source.

The purpose of this patch is to inform the user that in suspend to ram
mode, the wake source won't be taken into consideration.

Signed-off-by: Gregory CLEMENT <gregory.clement@free-electrons.com>