usb: xhci-mvebu: enable the COMPHY
This is just a minimal hack to have the xhci-mvebu driver get and
enable the COMPHY. The USB core already takes care of referencing and
enabling one PHY called "usb". But the Armada 375 actually references
two PHYs: the USB cluster and the COMPHY.

To be honest, I am not sure the USB cluster as a PHY was a good
idea. It isn't really a PHY. Maybe we should have simply used a
"syscon" and get the relevant drivers write what they need in this
register (though that would indeed not provide detection against
conflicts). Or it could be a pinctrl driver.

Or we keep it as a PHY, and we clean up this COMPHY handling: for now
it only enables the PHY: it never disables it nor it releases the
reference to it.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>