net: mvneta: remove data pointer usage from device_node structure
On year ago Rob Herring wanted to remove the data pointer from the
device_node structure[1]. The mvneta driver seemed to be the only one
which used (abused ?) it. However, the proposal of Rob to remove this
pointer from the driver introduced a regression and I proposed a
alternative way, but not as a proper patch.

Now here it is: In order to be able to remove the data pointer from the
device_node structure, we have to modify the way the BM resources are
shared between the mvneta port.

[1]https://www.spinics.net/lists/netdev/msg445197.html

Signed-off-by: Gregory CLEMENT <gregory.clement@bootlin.com>