net: mvpp2: use device_* APIs instead of of_* APIs
OF functions can be used only for the driver using DT.
As a preparation for introducing ACPI support in mvpp2
driver, use struct fwnode_handle in order to obtain
properties. For this purpose store its pointer in
struct device member of ports' net_device structures.
It had to be done, because actual struct device is common
for all ports and represent main PP2 controller part.

This patch replaces of_* function with device_* where possible.

Signed-off-by: Marcin Wojtas <mw@semihalf.com>