hugetlb, mbind: fall back to default policy if vma is NULL
Dan Carpenter has noticed that mbind migration callback (new_page) can get
a NULL vma pointer and choke on it inside alloc_huge_page_vma which relies
on the VMA to get the hstate.  We used to BUG_ON this case but the BUG_+ON
has been removed recently by "hugetlb, mempolicy: fix the mbind hugetlb
migration".

The proper way to handle this is to get the hstate from the migrated page
and rely on huge_node (resp.  get_vma_policy) do the right thing with null
VMA.  We are currently falling back to the default mempolicy in that case
which is in line what THP path is doing here.

Link: http://lkml.kernel.org/r/20180110104712.GR1732@dhcp22.suse.cz
Signed-off-by: Michal Hocko <mhocko@suse.com>
Reported-by: Dan Carpenter <dan.carpenter@oracle.com>
Cc: Naoya Horiguchi <n-horiguchi@ah.jp.nec.com>
Cc: Mike Kravetz <mike.kravetz@oracle.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Stephen Rothwell <sfr@canb.auug.org.au>