arm64: dts: marvell: move timer outside of AP810_NAME node
Since all APs are synchronizd with the AP0 timer, we only need to
describe the timers in the AP0. To eachieve this, we move the 'timer'
node outside of the AP810_NAME node. By doing this, all the 'timer'
nodes of the multiple instances of armada-ap810.dtsi will overwrite
each other, leaving only a single instance.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>