arm64: dts: marvell: prepare GIC description on AP810 for multi-AP
Every AP810 die includes a GIC Distributor. However, as a GIC-600
feature there is no need to initialize every distributor in a
multi-die system, since the GIC-600 broadcasts the distributor
configuration for other dies (this is done from AP0 in ATF firmware).

Therefore, we only need to describe the GIC for one AP. This commit
moves the GIC description outside of the AP810_NAME node, so that all
the definitions from the different AP instances override each other,
leaving only one.

However, as a sub-node of the GIC, we have the ITS, which has one
instance per AP. A new AP810_ITS_BASE must be defined by the code
including armada-ap810.dtsi, with the address of the ITS for the
current AP.

In addition, the GIC register addresses are changed, to use the "local
AP" range at 0xEC00_0000, which always points to the local AP GIC
registers.

The GIC base addresses are as follows:

    AP0:     	0xe8000000
    AP1:	0xe4000000
    AP2:	0xe0000000
    AP3:	0xdc000000

    local AP:	0xec000000

The GICD registers are at offset +0x3000000 from these base address,
so for example the local AP GICD registers are at 0xef000000.

The ITS registers are at 0x3040000 from this base address.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>