arm64: dts: marvell: make the msi-parent of the ICU configurable in armada-cp110
The ICU unit in the CP110 will have a different msi-parent when
integrated in the Armada 7K/8K SoCs or the upcoming 8KP SoC. This
commit therefore makes this property configurable, though a
pre-processor macro.

For now, it is always "gicp", but it will change with the introduction
of 8KP support.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>