arm64: dts: marvell: make the msi-parent of the PCIe configurable in armada-cp110
The PCIe controllers in the CP110 will have a different msi-parent
when integrated in the Armada 7K/8K SoCs or the upcoming 8KP SoC. This
commit therefore makes this property configurable, though a
pre-processor macro.

For now, it is always "gic_v2m0", but it will change with the
introduction of 8KP support. Also, notice that the ID of the PCIe
controller will be needed, even though it's not used for now.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>