arm64: dts: marvell: use intermediate macros for addresses in armada-cp110.dtsi
The armada-cp110.dtsi uses the ADDRESSIFY() macro in a few places to
turn a 12345678 string into a 0x12345678 address. For now, only the
low 32-bit of the address is used, so this ADDRESSIFY() macro is used
only once per adddress.

However, to enable using armada-cp110.dtsi for the Armada 8KP SoC, we
need to be able to use both the full 64-bit of the address, and using
ADDRESSIFY() twice for each address will make the DT hard to read. So
let's introduce intermediate macros for this.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>