arm64: dts: marvell: prepare for 64 bits addresses in armada-cp110.dtsi
The Armada 8KP will have registers and PCIe MEM/IO areas above the 4G
boundary, which requires to use full 64-bit addresses.

This commit prepares for that by introducing the upper32() and
lower32() macros in armada-common.dtsi, to separate the high 32 bits
and low 32 bits of an address. We then use them in armada-cp110.dtsi
in the appropriate places.

For the currently supported Armada 7K/8K, this commit is a no-op.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>