arm64: dts: marvell: move PCIe macros out of armada-cp110.dtsi
In order to allow the Armada 8KP to provide a completely different set
of macros for the PCIe controllers definitions, we move those
definitions out of armada-cp110.dtsi into armada-70x0.dtsi and
armada-80x0.dtsi.

It is worth mentioning that those definitions are for now duplicated
between armada-70x0.dtsi and armada-80x0.dtsi, because we don't have
any .dtsi common to Armada 7K/8K (but not 8KP). Perhaps this is
something we will want to change in the future.

The documentation at the top of armada-cp110.dtsi is updated to take
into account those changes.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>