arm64: dts: marvell: use define for CP110 pinctrl compatible
The Armada CP110 pinctrl compatible string needs to be different
between Armada 7K, Armada 8K first CP and second CP, and soon with the
Armada 8KP.

For now, we used DT inheritance to make sure the proper compatible
string is used depending on the SoC variant. However, on Armada 8KP,
we can have lots of CPs, and all of them will have the same pinctrl
compatible string. To avoid repeating this compatible string over and
over again, move it to the common armada-cp110.dtsi, and use a
pre-processor macro to define it, just like we're doing for lots of
over values in armada-cp110.dtsi already.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@bootlin.com>