net: mvpp2: Add RXAUI support
PPv2 is capable of 10Gbps rate using RXAUI (which is XGMII serialized
over 2 lanes, using BaseX encoding).

We need to use the correct XPCS configuration, and not expect the Idle
frames that aren't always emitted by the PHY.

We also need to set the correct Comphy configuration. RXAUI uses 2
comphy lanes, for now we pretend only one is used, and the comphy driver
does the proper lane assignment.

Signed-off-by: Maxime Chevallier <maxime.chevallier@bootlin.com>