net: mvpp2: Introduce VLAN_PRI HEK extraction
The classifier uses a Header Extracted Key for engine lookups, which is
composed of several fields extracted from the packer header, and
concatenated together.

One of these fields is the VLAN_PRI, which can only be extracted
whenever we are in a flow that contains tagged traffic. Add support for
that HEK field.

[TODO : Split out VLAN_PRI_QOS and the missing MAC_DA]
[TODO : Amended too fast, remove the whole macro redef from here]

Signed-off-by: Maxime Chevallier <maxime.chevallier@bootlin.com>