net: mvpp2: cls: Optimize vlan pri classification by using QoS tables
The C2 engine classifier has a pretty limited number of available
entries for classification. The 256 available entries are to be shared
between all ports on a PPv2, and then between each flow.

The C2 engine that's used for classification has some built-in tables
used for steering based on either the VLAN priority or the DSCP
priority, that allow to perform steering for all possible priorities
using only one C2 entry.

These tables are controlled through the "Action Tables" attributes of
the C2 entries. When using these tables for vlan priority steering, 64
tables of 8 entries each are available. One entry per flow and per port
is reserved for this so-called "QoS" steering.

These entries are however not real matches, and will always be considered as
"matching".

Signed-off-by: Maxime Chevallier <maxime.chevallier@bootlin.com>