net: mvpp2: divide fifo for dts-active ports only
Tx/Rx FIFO is a HW resource limited by total size, but shared
by all ports of same CP110 and impacting port-performance.
Do not divide the FIFO for ports which are not enabled in DTS,
so active ports could have more FIFO.

The active port mapping should be done in probe before FIFO-init.

Change-Id: I4f119f1a8a0b9462768c62b3c960908d5a005a77
Signed-off-by: Stefan Chulski <stefanc@marvell.com>